@@ -1,18 +1,16 @@
 apply plugin: 'com.android.library'
-apply plugin: 'maven'
-//apply plugin: 'com.github.dcendents.android-maven'
-//apply plugin: 'realm-android'
+apply plugin: 'com.github.dcendents.android-maven'
+
+group='com.github.philjay'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
-    // resourcePrefix 'mpcht'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
-        //noinspection MinSdkTooLow
-        minSdkVersion 9
-        targetSdkVersion 27
+        minSdkVersion 14
+        targetSdkVersion 28
         versionCode 3
-        versionName '3.0.3'
+        versionName '3.1.0'
     }
     buildTypes {
         release {
@@ -20,33 +18,14 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    lintOptions {
-        abortOnError false
-    }
     testOptions {
         unitTests.returnDefaultValues = true // this prevents "not mocked" error
     }
 }
 
-repositories {
-    maven {
-        url 'http://oss.jfrog.org/artifactory/oss-snapshot-local'
-    }
-}
-
 dependencies {
-    //provided 'io.realm:realm-android:0.87.5' // "optional" dependency to realm-database API
-    implementation 'com.android.support:support-annotations:27.1.1'
+    implementation 'androidx.annotation:annotation:1.0.0'
     testImplementation 'junit:junit:4.12'
-    testImplementation "org.mockito:mockito-core:1.10.19"
-}
-
-android.libraryVariants.all { variant ->
-    def name = variant.buildType.name
-    def task = project.tasks.create "jar${name.capitalize()}", Jar
-    task.dependsOn variant.javaCompiler
-    task.from variant.javaCompiler.destinationDir
-    artifacts.add('archives', task)
 }
 
 task sourcesJar(type: Jar) {
