@@ -1,8 +1,7 @@
-
 package com.github.mikephil.charting.animation;
 
 import android.animation.TimeInterpolator;
-import android.support.annotation.RequiresApi;
+import androidx.annotation.RequiresApi;
 
 /**
  * Easing options.
@@ -17,113 +16,6 @@ public class Easing {
     public interface EasingFunction extends TimeInterpolator {
         @Override
         float getInterpolation(float input);
-    }
-
-    /**
-     * Enum holding EasingOption constants
-     *
-     * @deprecated Use Easing.Linear instead of Easing.EasingOption.Linear
-     */
-    @Deprecated
-    public enum EasingOption {
-        Linear,
-        EaseInQuad,
-        EaseOutQuad,
-        EaseInOutQuad,
-        EaseInCubic,
-        EaseOutCubic,
-        EaseInOutCubic,
-        EaseInQuart,
-        EaseOutQuart,
-        EaseInOutQuart,
-        EaseInSine,
-        EaseOutSine,
-        EaseInOutSine,
-        EaseInExpo,
-        EaseOutExpo,
-        EaseInOutExpo,
-        EaseInCirc,
-        EaseOutCirc,
-        EaseInOutCirc,
-        EaseInElastic,
-        EaseOutElastic,
-        EaseInOutElastic,
-        EaseInBack,
-        EaseOutBack,
-        EaseInOutBack,
-        EaseInBounce,
-        EaseOutBounce,
-        EaseInOutBounce,
-    }
-
-    /**
-     * Returns the EasingFunction of the given EasingOption
-     *
-     * @param easing EasingOption to get
-     * @return EasingFunction
-     */
-    @SuppressWarnings("deprecation")
-    @Deprecated
-    public static EasingFunction getEasingFunctionFromOption(EasingOption easing) {
-        switch (easing) {
-            default:
-            case Linear:
-                return Easing.Linear;
-            case EaseInQuad:
-                return Easing.EaseInQuad;
-            case EaseOutQuad:
-                return Easing.EaseOutQuad;
-            case EaseInOutQuad:
-                return Easing.EaseInOutQuad;
-            case EaseInCubic:
-                return Easing.EaseInCubic;
-            case EaseOutCubic:
-                return Easing.EaseOutCubic;
-            case EaseInOutCubic:
-                return Easing.EaseInOutCubic;
-            case EaseInQuart:
-                return Easing.EaseInQuart;
-            case EaseOutQuart:
-                return Easing.EaseOutQuart;
-            case EaseInOutQuart:
-                return Easing.EaseInOutQuart;
-            case EaseInSine:
-                return Easing.EaseInSine;
-            case EaseOutSine:
-                return Easing.EaseOutSine;
-            case EaseInOutSine:
-                return Easing.EaseInOutSine;
-            case EaseInExpo:
-                return Easing.EaseInExpo;
-            case EaseOutExpo:
-                return Easing.EaseOutExpo;
-            case EaseInOutExpo:
-                return Easing.EaseInOutExpo;
-            case EaseInCirc:
-                return Easing.EaseInCirc;
-            case EaseOutCirc:
-                return Easing.EaseOutCirc;
-            case EaseInOutCirc:
-                return Easing.EaseInOutCirc;
-            case EaseInElastic:
-                return Easing.EaseInElastic;
-            case EaseOutElastic:
-                return Easing.EaseOutElastic;
-            case EaseInOutElastic:
-                return Easing.EaseInOutElastic;
-            case EaseInBack:
-                return Easing.EaseInBack;
-            case EaseOutBack:
-                return Easing.EaseOutBack;
-            case EaseInOutBack:
-                return Easing.EaseInOutBack;
-            case EaseInBounce:
-                return Easing.EaseInBounce;
-            case EaseOutBounce:
-                return Easing.EaseOutBounce;
-            case EaseInOutBounce:
-                return Easing.EaseInOutBounce;
-        }
     }
 
     private static final float DOUBLE_PI = 2f * (float) Math.PI;
