@@ -4,13 +4,12 @@ import com.github.mikephil.charting.data.Entry;
 import com.github.mikephil.charting.utils.ViewPortHandler;
 
 /**
- * Interface that allows custom formatting of all values inside the chart before they are
- * being drawn to the screen. Simply create your own formatting class and let
- * it implement IValueFormatter. Then override the getFormattedValue(...) method
- * and return whatever you want.
+ * Interface to format all values before they are drawn as labels.
  *
  * @author Philipp Jahoda
+ * @deprecated Extend {@link ValueFormatter} instead
  */
+@Deprecated
 public interface IValueFormatter
 {
 
@@ -24,6 +23,9 @@ public interface IValueFormatter
      * @param dataSetIndex    the index of the DataSet the entry in focus belongs to
      * @param viewPortHandler provides information about the current chart state (scale, translation, ...)
      * @return the formatted label ready for being drawn
+     *
+     * @deprecated Extend {@link ValueFormatter} and override an appropriate method
      */
+    @Deprecated
     String getFormattedValue(float value, Entry entry, int dataSetIndex, ViewPortHandler viewPortHandler);
 }
